requirements.txt
textfsm
# regexpro
# dlpro
pyyaml
